update User.jsx, SignupStep1.jsx, Privacy.jsx 수정 (개인정보 컴포넌트 수정)
